+++ springboot/distribusi.stok.barang/src/main/java/be/distribusi/stok/barang/repository/BarangRepository.java
@@ -8,6 +8,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -16,7 +17,7 @@ public interface BarangRepository extends JpaRepository<EntityBarang, Integer> {
     @Modifying
     @Transactional
     @Query(value = "INSERT INTO stok_barang (kode_barang, nama_barang, harga_beli, harga_jual, sisa_stok, stok_masuk, created_at) VALUES (?1, ?2, ?3, ?4, ?5, ?6, CURRENT_TIMESTAMP)", nativeQuery = true)
-    void insertStokBarang(String kodeBarang, String namaBarang, int hargaBeli, int hargaJual, int sisaStok, int stokMasuk);
+    void insertStokBarang(String kodeBarang, String namaBarang, BigDecimal hargaBeli, BigDecimal hargaJual, int sisaStok, int stokMasuk);
 
     // Query select by kode barang
     @Query(value = "SELECT CASE WHEN COUNT(*) > 0 THEN 1 ELSE 0 END FROM stok_barang WHERE kode_barang = ?1", nativeQuery = true)
